+++ android/app/src/main/kotlin/com/nomad/alarm/NomadAlarmApplication.kt
@@ -1,0 +1,10 @@
+package com.nomad.alarm
+
+import io.flutter.app.FlutterApplication
+
+class NomadAlarmApplication : FlutterApplication() {
+    override fun onCreate() {
+        super.onCreate()
+        GoogleMapsKeyHelper.loadAndApply(this)
+    }
+}